requirements.txt
asgiref==3.7.2
boto3==1.28.12
botocore==1.31.12
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==4.1.10
django-allauth==0.52.0
django-cors-headers==4.2.0
django-seed==0.3.1
django-storages==1.13.2
django-templated-mail==1.1.1
django-utils==0.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
djoser==2.2.2
dnspython==2.4.1
elastic-transport==8.11.0
elasticsearch==8.11.1
Faker==21.0.0
idna==3.4
jmespath==1.0.1
Markdown==3.4.4
mysqlclient==2.2.0
oauthlib==3.2.2
pdf2image==1.16.3
Pillow==10.0.0
psycopg==3.1.16
pycparser==2.21
PyJWT==2.8.0
pymongo==3.12.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
s3transfer==0.6.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.2.4
toposort==1.10
typing_extensions==4.7.1
urllib3==1.26.16
